Fixed checkmarks on dataelement and search.
