+++ src/app/shared/data-element-in-request/data-element-in-request.component.scss
@@ -38,6 +38,10 @@ section {
   margin-bottom: 0.5rem;
 }
 
+.fa-circle-check {
+  margin-left: 6px;
+}
+
 $triangle-center: 77.5%;
 $vertical-offset: 6px;
 $horizontal-offset: 8px;
